+++ atlasdb-config/src/main/java/com/palantir/atlasdb/http/AtlasDbHttpClients.java
@@ -15,6 +15,7 @@
  */
 package com.palantir.atlasdb.http;
 
+import java.net.ProxySelector;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
@@ -74,20 +75,56 @@ public final class AtlasDbHttpClients {
         return ret;
     }
 
-
     /**
-     * Constructs an HTTP-invoking dynamic proxy for the specified type that will cycle through the list of supplied
-     * endpoints after encountering an exception or connection failure, using the supplied SSL factory if it is
-     * present.
-     * <p>
-     * Failover will continue to cycle through the supplied endpoint list indefinitely.
+     * @deprecated please use {@link #createProxyWithFailover(Optional, Optional, Collection, Class)}, which requires
+     * you to specify the ProxySelector parameter.
      */
+    @Deprecated
     public static <T> T createProxyWithFailover(
             Optional<SSLSocketFactory> sslSocketFactory,
             Collection<String> endpointUris,
             Class<T> type) {
         return createProxyWithFailover(
                 sslSocketFactory,
+                Optional.empty(),
+                endpointUris,
+                type,
+                UserAgents.DEFAULT_USER_AGENT);
+    }
+
+    /**
+     * @deprecated please use {@link #createProxyWithFailover(Optional, Optional, Collection, Class, String)}, which
+     * requires you to specify the ProxySelector parameter.
+     */
+    @Deprecated
+    public static <T> T createProxyWithFailover(
+            Optional<SSLSocketFactory> sslSocketFactory,
+            Collection<String> endpointUris,
+            Class<T> type,
+            String userAgent) {
+        return createProxyWithFailover(
+                sslSocketFactory,
+                Optional.empty(),
+                endpointUris,
+                type,
+                userAgent);
+    }
+
+    /**
+     * Constructs an HTTP-invoking dynamic proxy for the specified type that will cycle through the list of supplied
+     * endpoints after encountering an exception or connection failure, using the supplied SSL factory if it is
+     * present. Also use the supplied the proxy selector to set the proxy on the clients if present.
+     * <p>
+     * Failover will continue to cycle through the supplied endpoint list indefinitely.
+     */
+    public static <T> T createProxyWithFailover(
+            Optional<SSLSocketFactory> sslSocketFactory,
+            Optional<ProxySelector> proxySelector,
+            Collection<String> endpointUris,
+            Class<T> type) {
+        return createProxyWithFailover(
+                sslSocketFactory,
+                proxySelector,
                 endpointUris,
                 type,
                 UserAgents.DEFAULT_USER_AGENT);
@@ -95,22 +132,26 @@ public final class AtlasDbHttpClients {
 
     public static <T> T createProxyWithFailover(
             Optional<SSLSocketFactory> sslSocketFactory,
+            Optional<ProxySelector> proxySelector,
             Collection<String> endpointUris,
             Class<T> type,
             String userAgent) {
         return AtlasDbMetrics.instrument(
                 type,
-                AtlasDbFeignTargetFactory.createProxyWithFailover(sslSocketFactory, endpointUris, type, userAgent),
+                AtlasDbFeignTargetFactory.createProxyWithFailover(
+                        sslSocketFactory, proxySelector, endpointUris, type, userAgent),
                 MetricRegistry.name(type));
     }
 
     @VisibleForTesting
     static <T> T createProxyWithQuickFailoverForTesting(
-            Optional<SSLSocketFactory> sslSocketFactory, Collection<String> endpointUris, Class<T> type) {
+            Optional<SSLSocketFactory> sslSocketFactory,
+            Optional<ProxySelector> proxySelector, Collection<String> endpointUris, Class<T> type) {
         return AtlasDbMetrics.instrument(
                 type,
                 AtlasDbFeignTargetFactory.createProxyWithFailover(
                         sslSocketFactory,
+                        proxySelector,
                         endpointUris,
                         QUICK_FEIGN_TIMEOUT_MILLIS,
                         QUICK_FEIGN_TIMEOUT_MILLIS,